+++ lib/widgets/paged_datatable/paged_datatable.dart
@@ -8,6 +8,7 @@ import '../../layout/constants/sizes.constant.dart';
 import '../buttons/cl_button.widget.dart';
 import '../buttons/cl_ghost_button.widget.dart';
 import '../buttons/cl_outline_button.widget.dart';
+import '../cl_popup_surface.widget.dart';
 import '../cl_shimmer.widget.dart';
 import '../cl_text_field.widget.dart';
 import '../cl_container.widget.dart';